IsRunning now allows jobs to continue if the endpoint doesn't have the API.
On the first upgrade with the new website, some sites may still be
running an older version of the service.  In those cases, they were
responding with 'Unable to tell if a job is running', which is true, but
will cause end-users to get confused and ask for help.

In this situation, we should eat the error, and continue with other
validations.